Fix issue with unhandled process events
It was possible to get an unhandled process event after a process
creation event.  On release builds, this unhandled event could be
interpreted by PresentMon as the process terminating (since the process
id could be the same due to the variable not being initialized, and the
image file name would be empty).  This could result in process names
being reported as "<error>".