Remove debug assert in PresentEvents to allow for clean event destruction
PresentData's destructor has asserts to catch incomplete events being destroyed. While this can be useful scenarios, it creates unexpected behaviors when these events are copied for diagnostic or record-keeping purposes.

My perspective is that this assert is better done in the debugger to keep the code clean and easy to use.

Alternatively, additional flags could help as a workaround, but isn't as clean.

Signed-off-by: Zuoming Shi <ZuomingShi@gmail.com>