Fix issue where delayed presents can be dequeued out-of-order.
In some cases where a present completes before all expected events are
observed, we defer its addition to mCompletePresentEvents until we see
the missing event.  When this happens, it is possible that a subsequent
present on the same swapchain is completed and added to
mCompletePresentEventss first, leading to presents from the same
swapchain being dequeued out of order.

PresentMon does not guarantee TimeInSeconds ordering throughout the
entire CSV, though it does require ordering between presents from the
same swapchain.

With this change, we move *all* completed presents into
mDeferredCompletions, keep them sorted per swapchain, and then only move
them into mCompletedPresents in order.

Test gold CSVs may need to be updated because this:
- may change ordering between different swapchains, and in particular
  between DWM and it's dependent presents
- may add or miss presents that happen near their process terminating,
  that weren't added/missed before this change.