Fix some issues due to submit sequence id collision.
Submit sequence id is only unique per context, but some events only
reference the submit sequence id which is ambiguous in some rare cases
(e.g., if there happens to be multiple queue submissions from different
contexts with the same sequence id).

This change modifies the mPresentBySubmitSequenceId map to key off of
both submit sequence id and context.  For events that do not reference
the context, the oldest present matching the sequence id is used.

Further, we remove presents from mPresentBySubmitSequence as early as
possible (e.g., as soon as they are known to be discarded) instead of
waiting for completion to further reduce the likelyhood of collisions.