Fix issue where full process path was reported.
When running presentmon as-admin, and when running presentmon before the
target application, the application process would be reported as
\Device\HarddiskVolumn...\...\Process.exe instead of Process.exe as
in other cases.

This change prunes off everything but the Process.exe for that case.

It also makes it more clear that ImageName is only valid for process
start events (since the property on ProcessStop is not reliable).

Addresses issue #157