Fix issue where only part of a process name was being compared
The issue was that when a process was identified, it was only being
compared to the first n characters of the -exclude and -process_name
names, where n was the identified process name's length after any
directory and .exe extension had been removed.

e.g., using -exclude code.exe would also cause cod.exe to be excluded
(through, not vice versa).

See issue #172.