Fix Composed display tracking on latest windows OS
Fixes an issue observed where DWM may present without API present
events. In this case, the PresentEvent is created when the Flip event is
observed.  However, because there is no PresentStop event, it isn't
removed from mPresentByThreadId and the next Flip event will look it up
and think it is a lost present.